(MODULES-9446) Store type not set properly
The keytool.rb provider does not set the keytool type properly.
Prior to this change the comparison at line 217 of the provider was
trying to compare a variable of type symbol to a string literal. The
failure of this comparison caused the `-storetype` parameter to the
keytool command to be omitted in error.

This change fixes the comparison types to fix the error.